refactor(window): to make it more readable
