+++ flock/Cargo.toml
@@ -26,7 +26,7 @@ fixedbitset = { version = "0.4.0", optional = true }
 futures = "0.3.12"
 glob = { version = "0.3", optional = true }
 hashbrown = "0.12"
-hashring = "0.2.0"
+hashring = { git = "https://github.com/flock-lab/hashring-rs", branch = "flock" }
 humantime = "2.1.0"
 indoc = "1.0.3"
 itertools = "0.10.0"
